requirements.txt
boto3==1.4.4
botocore==1.5.72
click==6.7
docutils==0.13.1
jmespath==0.9.3
progressbar2==3.30.2
python-dateutil==2.6.0
python-utils==2.1.0
s3transfer==0.1.10
six==1.10.0
